More bug fixes
No longer setting fortran flags as a variable in efit since it cannot
fit on a single line and thus breaks clang compilers.

Resetting all reals explicitly to 8 bytes unless they are indicated to
be conversion variables.  Compilers were catching subroutine calls with
inconsitent types, so any conversions need to be handled more carefully.

+++ efit/flux_current.F90
@@ -578,7 +578,7 @@
 
       dimension pds(6)
       integer*4 iii
-      real :: zmaxis_last = 0.0
+      real*8 :: zmaxis_last = 0.0
       data isplit/8/,psitol/1.0e-04_dp/,cdum/1.0/
       save xguess, yguess, xltrac, radbou
 !
