+++ efit/msels_data.f90
@@ -49,7 +49,7 @@
       character*3 synmls
       integer*4 i,j,count
       integer*4 ishot, ktime, icmls, iermls(ktime), l_iermls
-      real*4 avem, atime(ktime), bbmls(ktime), sigbmls(ktime),        &
+      real*8 avem, atime(ktime), bbmls(ktime), sigbmls(ktime),        &
              rrmls(ktime), zzmls(ktime),                              &
              L1mls(ktime), L2mls(ktime), L4mls(ktime),                &
              epotpmls(ktime), sigepmls(ktime)
